fixing compiler warnings

See merge request njoy/ENDFtk!54

+++ src/ENDFtk/section/1/458/PolynomialComponents/src/print.hpp
@@ -13,7 +13,7 @@ void print( OutputIterator& it, int MAT, int MF, int MT ) const {
 
   std::vector< double > values;
   values.reserve( 18 * ( this->order() + 1 ) );
-  for ( std::size_t i = 0; i < this->order() + 1; ++i ) {
+  for ( std::size_t i = 0; i < static_cast< std::size_t >( this->order() ) + 1; ++i ) {
 
     for ( std::size_t j = 0; j < 9; ++j ) {
